Add build android. Add helpers file and refactor
