+++ packages/cli/src/commands/build-android.ts
@@ -1,0 +1,16 @@
+import {BuildPlatform} from '../types';
+import {createBuildDir, bundleJS, getBuildDir} from '../tools/helpers';
+
+export default function buildAndroid(args: BuildPlatform) {
+  const rootDir = process.cwd();
+  const buildDir = getBuildDir(rootDir);
+
+  createBuildDir(buildDir);
+
+  bundleJS({
+    ...args,
+    platform: 'android',
+    rootDir,
+    buildDir,
+  });
+}
