+++ packages/cli/src/types/index.ts
@@ -1,4 +1,12 @@
-export type BuildArtifact = {
+export type BuildPlatform = {
   entryFile: string;
   useNpm?: boolean;
 };
+
+export type Platform = 'ios' | 'android';
+
+export type BundleJSArgs = {
+  platform: Platform;
+  buildDir: string;
+  rootDir: string;
+} & BuildPlatform;
